+++ modules/aws-eks/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.1.1](https://github.com/prefapp/tfm/compare/aws-eks-v1.1.0...aws-eks-v1.1.1) (2025-04-24)
+
+
+### Bug Fixes
+
+* Update role_assignment.tf ([7592841](https://github.com/prefapp/tfm/commit/75928419415d74de12d2d38a602df7aa703c860e))
+
+## [1.1.0](https://github.com/prefapp/tfm/compare/aws-eks-v1.0.5...aws-eks-v1.1.0) (2025-04-24)
+
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
 ## [1.0.5](https://github.com/prefapp/tfm/compare/aws-eks-v1.0.4...aws-eks-v1.0.5) (2025-04-01)