+++ modules/azure-redis-cache/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [0.2.2](https://github.com/prefapp/tfm/compare/azure-redis-cache-v0.2.1...azure-redis-cache-v0.2.2) (2025-05-05)
+
+
+### Bug Fixes
+
+* move resources previously created ([#643](https://github.com/prefapp/tfm/issues/643)) ([8a46be8](https://github.com/prefapp/tfm/commit/8a46be82953952f147cb3ae228cacb0703c4c63a))
+
+## [0.2.1](https://github.com/prefapp/tfm/compare/azure-redis-cache-v0.2.0...azure-redis-cache-v0.2.1) (2025-04-24)
+
+
+### Bug Fixes
+
+* Update role_assignment.tf ([7592841](https://github.com/prefapp/tfm/commit/75928419415d74de12d2d38a602df7aa703c860e))
+
+## [0.2.0](https://github.com/prefapp/tfm/compare/azure-redis-cache-v0.1.0...azure-redis-cache-v0.2.0) (2025-04-24)
+
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
+## [0.2.0](https://github.com/prefapp/tfm/compare/azure-redis-cache-v0.1.0...azure-redis-cache-v0.2.0) (2025-04-24)
+
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
 ## 0.1.0 (2025-03-21)