+++ modules/cloudamqp-cluster/CHANGELOG.md
@@ -1,5 +1,112 @@
 # Changelog
 
+## [1.1.1](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v1.1.0...cloudamqp-cluster-v1.1.1) (2025-04-24)
+
+
+### Bug Fixes
+
+* Update role_assignment.tf ([7592841](https://github.com/prefapp/tfm/commit/75928419415d74de12d2d38a602df7aa703c860e))
+
+## [1.1.0](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v1.0.0...cloudamqp-cluster-v1.1.0) (2025-04-24)
+
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
+
+### Bug Fixes
+
+* Update role_assignment.tf ([7592841](https://github.com/prefapp/tfm/commit/75928419415d74de12d2d38a602df7aa703c860e))
+* value ([5cacf36](https://github.com/prefapp/tfm/commit/5cacf36592574c95eb252629054c0033338d6ee6))
+
+## [1.1.0](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v1.0.0...cloudamqp-cluster-v1.1.0) (2025-04-24)
+
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
+
+### Bug Fixes
+
+* Update role_assignment.tf ([7592841](https://github.com/prefapp/tfm/commit/75928419415d74de12d2d38a602df7aa703c860e))
+* value ([5cacf36](https://github.com/prefapp/tfm/commit/5cacf36592574c95eb252629054c0033338d6ee6))
+
+## [1.1.0](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v1.0.0...cloudamqp-cluster-v1.1.0) (2025-04-24)
+
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
+
+### Bug Fixes
+
+* value ([5cacf36](https://github.com/prefapp/tfm/commit/5cacf36592574c95eb252629054c0033338d6ee6))
+
+## [1.0.0](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v0.1.0...cloudamqp-cluster-v1.0.0) (2025-04-24)
+
+
+### ⚠ BREAKING CHANGES
+
+* vpc_id reference and vpc associated into cloudamp-cluster module ([#505](https://github.com/prefapp/tfm/issues/505))
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
+
+### Bug Fixes
+
+* remove sensitive api_key output ([#455](https://github.com/prefapp/tfm/issues/455)) ([cbb7d72](https://github.com/prefapp/tfm/commit/cbb7d728cc1a0097f4b7c8c584080dad5f4b5a1c))
+* value ([5cacf36](https://github.com/prefapp/tfm/commit/5cacf36592574c95eb252629054c0033338d6ee6))
+* vpc_id reference and vpc associated into cloudamp-cluster module ([#505](https://github.com/prefapp/tfm/issues/505)) ([91d324b](https://github.com/prefapp/tfm/commit/91d324b725a1d9bf3ef15eb5e7f2816be619fcac))
+
+## 0.1.0 (2025-04-24)
+
+
+### ⚠ BREAKING CHANGES
+
+* vpc_id reference and vpc associated into cloudamp-cluster module ([#505](https://github.com/prefapp/tfm/issues/505))
+
+### Features
+
+* Adding cloudamqp-cluster module ([#379](https://github.com/prefapp/tfm/issues/379)) ([7d95d7c](https://github.com/prefapp/tfm/commit/7d95d7c6a59e53c1a6807a3821d50fa4c0be438b))
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
+
+### Bug Fixes
+
+* remove sensitive api_key output ([#455](https://github.com/prefapp/tfm/issues/455)) ([cbb7d72](https://github.com/prefapp/tfm/commit/cbb7d728cc1a0097f4b7c8c584080dad5f4b5a1c))
+* value ([5cacf36](https://github.com/prefapp/tfm/commit/5cacf36592574c95eb252629054c0033338d6ee6))
+* vpc_id reference and vpc associated into cloudamp-cluster module ([#505](https://github.com/prefapp/tfm/issues/505)) ([91d324b](https://github.com/prefapp/tfm/commit/91d324b725a1d9bf3ef15eb5e7f2816be619fcac))
+
+## [1.1.0](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v1.0.1...cloudamqp-cluster-v1.1.0) (2025-04-24)
+
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
+## [1.0.1](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v1.0.0...cloudamqp-cluster-v1.0.1) (2025-04-22)
+
+
+### Bug Fixes
+
+* value ([5cacf36](https://github.com/prefapp/tfm/commit/5cacf36592574c95eb252629054c0033338d6ee6))
+
+## [1.0.0](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v0.1.1...cloudamqp-cluster-v1.0.0) (2025-04-22)
+
+
+### ⚠ BREAKING CHANGES
+
+* vpc_id reference and vpc associated into cloudamp-cluster module ([#505](https://github.com/prefapp/tfm/issues/505))
+
+### Bug Fixes
+
+* vpc_id reference and vpc associated into cloudamp-cluster module ([#505](https://github.com/prefapp/tfm/issues/505)) ([91d324b](https://github.com/prefapp/tfm/commit/91d324b725a1d9bf3ef15eb5e7f2816be619fcac))
+
 ## [0.1.1](https://github.com/prefapp/tfm/compare/cloudamqp-cluster-v0.1.0...cloudamqp-cluster-v0.1.1) (2025-03-04)