+++ modules/mongodb-atlas-project/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.4.1](https://github.com/prefapp/tfm/compare/mongodb-atlas-project-v1.4.0...mongodb-atlas-project-v1.4.1) (2025-04-24)
+
+
+### Bug Fixes
+
+* Update role_assignment.tf ([7592841](https://github.com/prefapp/tfm/commit/75928419415d74de12d2d38a602df7aa703c860e))
+
+## [1.4.1](https://github.com/prefapp/tfm/compare/mongodb-atlas-project-v1.4.0...mongodb-atlas-project-v1.4.1) (2025-04-24)
+
+
+### Bug Fixes
+
+* Update role_assignment.tf ([7592841](https://github.com/prefapp/tfm/commit/75928419415d74de12d2d38a602df7aa703c860e))
+
+## [1.4.0](https://github.com/prefapp/tfm/compare/mongodb-atlas-project-v1.3.1...mongodb-atlas-project-v1.4.0) (2025-04-24)
+
+
+### Features
+
+* Update README.md ([a7dfb55](https://github.com/prefapp/tfm/commit/a7dfb55b83447cf3ef08d168ab756e791f322e7a))
+
 ## [1.3.1](https://github.com/prefapp/tfm/compare/mongodb-atlas-project-v1.3.0...mongodb-atlas-project-v1.3.1) (2025-01-21)